Revert "Revert "Merge pull request #5 from primitivefinance/release/hotfix-missing-interfaces""
This reverts commit 83472bad2fdab4b13680faafc9259ec13b9444e8.

+++ packages/primitive-contracts/contracts/Primitives.sol
@@ -1,13 +1,12 @@
 // SPDX-License-Identifier: MIT
-
 pragma solidity ^0.6.2;
 
 library Primitives {
     struct Option {
         address underlyingToken;
         address strikeToken;
-        uint base;
-        uint quote;
-        uint expiry;
+        uint256 base;
+        uint256 quote;
+        uint256 expiry;
     }
 }
